feat (agent): add callback for type 0 & 1 messages being sent to the agent
with verification in type 1 with the secret again, and if the type is 0
then we just return because that's a bit strange. because if we send out
the broadcast the handler catches it, an agent shouldnt need to catch
it? im not sure, need to fix the bug i talked about